validate that users in connection must be vetted only when connections are established among non-ketchy users.
A bug was discovered by Arman when he tried to log in and we tried to connect him with Ketchy. Because Arman has not been vetted, his connection to Ketchy was rejected. Because Ketchy is a special entity, we would like to allow Arman to connect with Ketchy and only perfom this validation when 2 non-Ketchy users are trying to be connected.